fix caching issues
* ensure changed group names do not lead to broken links
* ensure redirect path in login form does not get cached
* include version in wiki cache key
  just in case two updates happen in the same second
* turn on caching in tests - this way you find these issues

Cache keys of page listings need to change if the owners name changes.
Also there is no need to include a duplicate timestamp (one for the
page.update_at and one for the formatted date).

So now we only use the formatted date and the page.uri - which includes
the owner_name and the page name.