paranoid email digest - still needs to be tweaked
So far we used to send one digest per changed page per day.
I think we should only send one digest mail per day and list the pages
in it. That's the point of a digest.
However with paranoid emails this will look fairly boring...
A page you are watching has been modified 4 times.
Another page you are watching has been modified 1 time.
...

So instead we just notify the user about the total of pages changed and
point them to their dashboard for updates.

Now the translations won't match what is going on. So the strings
need to be updated and we will need new translations.