define post_path in Me::NoticesController
show tries to redirect to the noticable_path - for PostStarred Notices this is
a post_path.

We don't have a  post_path yet. Mostly due to the way posts are attached to pages
and relationships via discussions. The proper fix would be to turn them into
comments (for pages) and messages (between two users) and then have non
nested urls for them. ... but that is something for later. For now this will fix the
500 response and replace it with a meaningful redirect.