bugfix: sortable image list after file upload
The url for the sortable_element was wrong. It would also render the collection of images in a wrong order.

We now work around the latter by using @page.images if it's available
(gallery) rather than @page.assets (other page types). images are sorted
by position based on showing wheras assets are not sorted right now.

The longer term fix for this will be to add a position field to the
assets and remove the showings table all together (we don't allow
reusing images across galeries for a long time now.)