only create a symlink for public asset dirs if needed
Before we tested for the existance of the file and tried to symlink the
directory if it did not exist. Then we would redirect to the same url
assuming this fixed things.

Turns out that if there is a different issue such as the file missing -
not the symlink to the dir this does not fix thigns but leads to endless
redirects.

So now we only attempt to fix things by symlinking if there was no symlink
yet.