upgrade: use utf8-cleaner to prevent 400
We still have some links in wikis etc. that use the old non utf-8
based % encoding. Some of these are invalid these days.
But we can still identify the assets they point at etc. so we just
drop them. We used to do this in the controller but now the
error is already raised in the middleware. So instead we use a the
utf-8 cleaner gem.