cleanup: turn Site into plain ruby class and simplify
We used to allow a LOT of configuration via sites and Conf.
This developed out of the need to create working instances for
different interests. Now this is no more the case and we can
finally drop the associated complexity.

The need to persist site settings in the database is also gone
since the is no UI for changing them anymore. Instead they can
now be applied via configuration files only.

While i ripped out a lot of complexity and brittle code i kept
a basic destinction. Some config settings are accessed via
  Conf
others in the controllers and views via
  current_site

This will still allow us to customize ui and to some extend
behaviour based on the domain of a request. It will NOT serve
however to run two completely different sites on the same install
though.