bugfix: remove caching of page entities
We would also have to expire the cache every time a link or a name changes.
This boils down to
  cache @group, @group.committees.count, @group.committees.max(:updated_at), ... do
Currently rendering the corresponding lists is not really slow - 20 ms. for all the
entities of animals.

So if we have to query every association for it's size and the max updated at
we probably won't gain much and this kind of caching is error prone.