Revert "bugfix: set correct path for assets manifest"
This reverts commit 95f4d408426a2c3c3e12451d686ddc7c0bb66dcc.

The manifest nowerdays is specified as a filename - not as a path.
So this caused problems in production.

Somehow the original problem in production is gone never the less:
The icon_png stylesheet is not referrenced with an integrity hash
in the html head. No idea what caused sprockets to kick in in the end.