refactor: reduce complexity of page_contexts_controller
Removed logic for identifying pages without context. It looks like
this is broken in production and i am not aware of it being used in
the wild. We also have no test covering it and it adds a significant
overhead because it required the Page::Finder to rely on
Page.paginate_by_path which in turn needed the path finder options
from the controller.

If we wanted logic like this to magically turn up the right page when
typing in a wrong url we might add it to the 404 page as suggestions.