refactor: minor reorganization of TagSuggestions
I would like to move Page.tags_for_group(page.owner, user) into the
TagSuggestions class as that seems a more natural fit. However
it is also used in a search filter. This is a bit of preparation
for changing the TagSuggestions so an equivalent to tags_for_group
can easily be exposed.