fix: preserve base class model_names
uhh ohh.... was overwriting methods in the singleton class of User
and Group instead of creating a copy first. This one has been
bugging us on live systems for quite a while. Now that @dgt created
a test that failed deterministically (with the right salt) I was
finally able to debug. :)