cleanup: last alias_method_chain
:notes:
we don't need no alias_method.
we don't need no prepend module.

no method chaining in the class def.

hey monkey - patch this code alone!

all in all it's just a redefined method call.
:notes:

in other words:
since all of the methods we expect are defined in modules (as they are
permissions) they are defined in ancestors of the actual controller. So
we can overwrite them and call super with no need for prepend or
alias_method_chain.