fix: rerender the full wiki when saving section
This was broken lately by f5b13098c543d2a384f9759cf9b578174f2dab5b.
Instead of replacing the full wiki it would now be rendered into the
section.

We need to rerender the full wiki in order for changes to the table
of contents to take place but also in case we entirely remove a section
and so on.