fix: #133 invalid page values cause 500s
We have a lot of scans recently and since php loves ?page=...
people try that. Each time they do the server returns a 500
which clutters our logs and burries actual errors.

Let us log this with a specific warning instead and carry
on. In most cases this will instead lead to a 404 because
the dispatch controller cannot find the path they are trying.