check if a key is expired and deal with broken keys
added test, moved keys to fixtures
still a FIXME open: mail-gpg does not throw errors when dealing with a broken/expired
key. Instead it falls back to keys found in the keyring (=> we should provide a fix
for mail-gpg
for now we create a new keyring for each page history. we do not do this
for the key upload mailer yet