introduce Message and MessagePolicy
This way we can cleanup PostPermission.

The remaining problem was checking the create? policy.
In order to do that we would like to know the recipient.

Retrieving the recipient from the post is a pain.
So instead we now have a plain ruby class that wraps
and exposes the parameters required for message creation.
Checking the Policy based on it is trivial.

Next we can probably move the complex Message creation
logic from the user model into the new class.