cleanup: permissions in survey page
survey permission system is fairly complex.
Changed the controller permission checks to be pundit based.

Moved the more fine grained permission checks in the views
to the SurveyPermissionHelper.

We only display existing surveys. We already dropped the ability
to create or modify them.

This commit also removes all the now unused permission calls.