cleanup: people permission
replaced may_pester with a pundit based call of
  may_create? message
with a message object just created for that.

We might actually make use of that message object
to create the path to link to.
However this will require expanding Message a bit
and making sure our routes and the routes generated by message match.
... a bit too much for now.