pundit: verify_authorized in more controllers
verify_authorized in the ApplicationController is too generic.
I tried to push it up the inheritance chain where possible.
For example Person::BaseController now verifies authorization.
This way it is not required in Person::FriendRequestsController.

The authorization itself now happens in the actions more often.
In particular the actions in the Common::Requests module
now explicitly skip the authentication.

I am pondering adding a RequestPolicy that basically calls
:may_approve? on the request in question for updates.

It would be redundant with the model level authentication in requests.
But at the same time it would be consistent with how we authenticate
in other places.
Redundant authentication is not necessarily a bad thing either.