test: clear key cache between tests
One test grants the permission
while the other one tests default behaviour.

If they ran in wrong order permission would be cached in User key_cache

We should get rid of that cache at some point.
But that is unrelated to the current work on using pundit.
So leaving it for later and adding a TODO.