pundit: clean up group and membership policy a bit
We still have a bunch of may_create_... functions.
But it is getting better.

Adds a @membership instance var to all Group::BaseController subclasses.
This holds either the current_users membership record
or a new_record that serves as a null object.

Note that in order to check if the current_user actually is a member
you need to check @membership.persisted?
- the fact that @membership is present is not enough.