fix: prevent 500 on invalid format urls
during scans we often see requests for
  /test.xml
and similar.

Since a test group exists this request will trigger the Group::HomeController
and attempt to render an xml version.
This leads to a  template not found exception resulting in a 500 response.

Instead we now only accept html requests to the context_urls (the ones without a prefix).
This should deal with most of these requests.

We might want to add format constraints to all routes
 - this would result in blocking requests