No access through user_participation when creating group page
we do not need user_participations, as long as the user has access through the group_participation.
on group page creation, we only add a user_participation for pages in public groups of which the user
is not a member.

the tests now show the different owner handling
- the page creator will become the owner whenever the owner is invalid (a
group without access etc.)