Prepare for utf8mb4: shrink size of indexed fields
to make database indexes fit after conversion to full utf8, the db forces
us to shrink the size of the varchar fields which are part of those indexes.
before doing so, we had to make sure that the content in this fields does not
exceed 191 characters. we used a rake task to shrink long legacy tags.