Switch to full unicode (utf8mb4) to display emojis
switch to utf8mb4 which can store all unicode code points including
emoticons (utf8 which we used before is an alias for utf8mb3 which only
stores a maximum of three bytes).

the rake task cg:convert_to_unicode converts the database into utf8mb4.
it also sets a binary collation for our tags table, because we
want to distinguish between olé and ole. there is no test for this
feature (because it would fail).
the schema_migrations table has to be excluded, because it has
indexes which are to long

the emoji tests work without database conversion, because database
connection uses utf8mb4.