ci: do not run cleanup task
It is pretty meaningless since tests run in transactions.
At the end of every test the database is reset to the state
it had right after loading the fixtures.

Therefore all the invalid records are already present in the fixtures
except for the page terms.

PageTerms use a different kind of table that is incompatible
with transactional fixtures. Therefore we have some page terms
left after running the tests that do not have corresponding pages.

This however is a test artifact and not telling us anything
other than that we create pages during tests and clean them up
but leave the page terms around.