handle 404 ourself rather than with exception app
The exceoption app middleware runs after the cookie handler.
Therefore sessions are not available to it
which results in the login form becoming invalid.

In addition exceptions handled by the exception app will still be logged.
This is cluttering our logs quite a bit.
We do not need traces for everything that cannot be found or accessed.

This commit brings back the render_not_found method in Controllers.

For the Dispatch controllers it dispatches to the exceptions controller
directly rather than using it via the exception app.

Keep the assert_not_found helper with a block
even though it is technically not needed anymore.
Before this used assert_raises.
Since we do not raise anymore this is not needed and we now use assert_response.
Keeping the same style will allow us to roll this back more easliy
and reduces the number of places we need to change.