fix: render 401 ourselves
Just like 404s these were rendered by the exception app before.
Since the exception app runs as a rake middleware
after handling cookies and sessions
cookies and sessions did not work on the error pages.

This was a problem for 401s
because we render them mostly when people need to login.
However the lack of cookies and sessions broke the login form.