test: use orange in testing peers rather than red
adding and removing red to and from animals
resulted in red getting removed from the first page.
This in turn updated the page_terms
which were inconsistent with user participations after the next fixture load.

This resulted in the page_search test failing occasionally

orange is a great drop in replacement for red in the social test.