cleanup: custom javascript and stylesheet handling
We use the assets pipeline for js and css for a while now.
We also moved away from assets as_needed by specific controllers.
All the js and all the css are send all the time to allow for better caching.

Remove the remains of the per controller asset handling