feature: prevent creation of spam posts
We have seen a lot of spam comments on public pages recently.
They insert links - probably to increase search engine ranking of the linked sites.

In order to prevent this we disallow comments with links on public pages
for users who have no other access to the page than it being public.